Abstracts button dropdown and card
